+++ pkg/psmdb/service_test.go
@@ -1,35 +1,132 @@
 package psmdb
 
 import (
+	"context"
 	"testing"
 
+	"github.com/pkg/errors"
 	"github.com/stretchr/testify/assert"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/util/intstr"
+	"k8s.io/client-go/kubernetes/scheme"
 	"k8s.io/utils/ptr"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+	"sigs.k8s.io/controller-runtime/pkg/client/fake"
 
 	api "github.com/percona/percona-server-mongodb-operator/pkg/apis/psmdb/v1"
 	"github.com/percona/percona-server-mongodb-operator/version"
 )
 
+func TestMongosHost(t *testing.T) {
+	ctx := context.Background()
+
+	pod := &corev1.Pod{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      "mongos-0",
+			Namespace: "default",
+		},
+	}
+
+	tests := map[string]struct {
+		init          func(cl client.Client)
+		expectedHost  string
+		expectedError error
+	}{
+		"service not found": {
+			init:         func(cl client.Client) {},
+			expectedHost: "",
+		},
+		"clusterip service type": {
+			init: func(cl client.Client) {
+				svc := &corev1.Service{
+					ObjectMeta: metav1.ObjectMeta{
+						Name:      "test-cluster-mongos",
+						Namespace: "default",
+					},
+					Spec: corev1.ServiceSpec{
+						Type: corev1.ServiceTypeClusterIP,
+						Ports: []corev1.ServicePort{
+							{
+								Name: "mongos",
+								Port: 27018,
+							},
+							{
+								Name: "random",
+								Port: 12345,
+							},
+						},
+					},
+				}
+				assert.NoError(t, cl.Create(ctx, svc))
+			},
+			expectedHost: "test-cluster-mongos.default.svc.cluster.local:27018",
+		},
+		"err: clusterip service type and port not found": {
+			init: func(cl client.Client) {
+				svc := &corev1.Service{
+					ObjectMeta: metav1.ObjectMeta{
+						Name:      "test-cluster-mongos",
+						Namespace: "default",
+					},
+					Spec: corev1.ServiceSpec{
+						Type: corev1.ServiceTypeClusterIP,
+						Ports: []corev1.ServicePort{
+							{
+								Name: "random",
+								Port: 12345,
+							},
+						},
+					},
+				}
+				assert.NoError(t, cl.Create(ctx, svc))
+			},
+			expectedError: errors.New("mongos port not found in service"),
+		},
+	}
+	for name, tt := range tests {
+		t.Run(name, func(t *testing.T) {
+			s := scheme.Scheme
+			cl := fake.NewClientBuilder().WithScheme(s).Build()
+			assert.NotNil(t, cl)
+			tt.init(cl)
+
+			cr := &api.PerconaServerMongoDB{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test-cluster",
+					Namespace: "default",
+				},
+				Spec: api.PerconaServerMongoDBSpec{
+					ClusterServiceDNSSuffix: "svc.cluster.local",
+					Sharding: api.Sharding{
+						Mongos: &api.MongosSpec{
+							Expose: api.MongosExpose{
+								ServicePerPod: false,
+							},
+						},
+					},
+				},
+			}
+
+			host, err := MongosHost(ctx, cl, cr, pod, false)
+			if tt.expectedError != nil {
+				assert.Empty(t, host)
+				assert.EqualError(t, err, tt.expectedError.Error())
+				return
+			}
+			assert.Equal(t, tt.expectedHost, host)
+			assert.NoError(t, err)
+		})
+	}
+}
+
 func TestExternalService(t *testing.T) {
 	tests := map[string]struct {
-		cr          *api.PerconaServerMongoDB
 		replset     *api.ReplsetSpec
 		podName     string
 		expectedSvc *corev1.Service
 	}{
 		"ClusterIP": {
-			cr: &api.PerconaServerMongoDB{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      "test-cr",
-					Namespace: "test-ns",
-				},
-				Spec: api.PerconaServerMongoDBSpec{
-					CRVersion: version.Version,
-				},
-			},
 			replset: &api.ReplsetSpec{
 				Name: "rs0",
 				Expose: api.ExposeTogglable{
@@ -82,15 +179,6 @@ func TestExternalService(t *testing.T) {
 			},
 		},
 		"NodePort": {
-			cr: &api.PerconaServerMongoDB{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      "test-cr",
-					Namespace: "test-ns",
-				},
-				Spec: api.PerconaServerMongoDBSpec{
-					CRVersion: version.Version,
-				},
-			},
 			replset: &api.ReplsetSpec{
 				Name: "rs0",
 				Expose: api.ExposeTogglable{
@@ -144,15 +232,6 @@ func TestExternalService(t *testing.T) {
 			},
 		},
 		"LoadBalancer": {
-			cr: &api.PerconaServerMongoDB{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      "test-cr",
-					Namespace: "test-ns",
-				},
-				Spec: api.PerconaServerMongoDBSpec{
-					CRVersion: version.Version,
-				},
-			},
 			replset: &api.ReplsetSpec{
 				Name: "rs0",
 				Expose: api.ExposeTogglable{
@@ -212,7 +291,16 @@ func TestExternalService(t *testing.T) {
 	}
 	for name, tt := range tests {
 		t.Run(name, func(t *testing.T) {
-			svc := ExternalService(tt.cr, tt.replset, tt.podName)
+			cr := &api.PerconaServerMongoDB{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test-cr",
+					Namespace: "test-ns",
+				},
+				Spec: api.PerconaServerMongoDBSpec{
+					CRVersion: version.Version,
+				},
+			}
+			svc := ExternalService(cr, tt.replset, tt.podName)
 			assert.Equal(t, tt.expectedSvc, svc)
 		})
 	}
